Setup drep command group
