+++ cardano-cli/test/cardano-cli-golden/files/golden/help/shelley_governance.cli
@@ -1,9 +1,4 @@
-Usage: cardano-cli shelley governance 
-                                        ( registration-certificate
-                                        | delegation-certificate
-                                        | create-mir-certificate
-                                        | action
-                                        )
+Usage: cardano-cli shelley governance (create-mir-certificate | action | drep)
 
   Era-based governance commands
 
@@ -11,8 +6,7 @@ Available options:
   -h,--help                Show this help text
 
 Available commands:
-  registration-certificate Create a registration certificate.
-  delegation-certificate   Delegation certificate creation.
   create-mir-certificate   Create an MIR (Move Instantaneous Rewards)
                            certificate
   action                   Governance action commands.
+  drep                     DRep member commands.
